alx_travel_app/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
django-cors-headers>=4.2.0
django-environ>=0.11.2
mysqlclient>=2.2.4
celery>=5.3.6
django-celery-results>=2.5.1
django-celery-beat>=2.6.0
pytz>=2023.3
redis>=5.0.1
requests>=2.31.0
